按实际返回的 JSON 计算行情读字节上限，分页偏移改用 long，空密码启动即失败。
cursor-tracy-mbp

+++ common/src/main/java/world/willfrog/alphafrogmicro/common/datasource/MarketReadPageResult.java
@@ -5,5 +5,5 @@ import java.util.List;
 /**
  * 一页行情读结果。{@code hasMore=true} 表示调用方用 limit+1 探到后续页，不是静默截断。
  */
-public record MarketReadPageResult<T>(List<T> rows, int offset, int limit, boolean hasMore) {
+public record MarketReadPageResult<T>(List<T> rows, long offset, int limit, boolean hasMore) {
 }
